requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.0
Brotli==1.0.9
CacheControl==0.12.11
cachetools==5.3.1
cachy==0.3.0
cairocffi==1.4.0
CairoSVG==2.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
crashtest==0.3.1
cryptography==41.0.1
cssselect2==0.7.0
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.1.0
Django==4.1.3
django-colorfield==0.8.0
django-cors-headers==3.14.0
django-crontab==0.7.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-jazzmin==2.6.0
django-light==0.1.0.post3
django-mathfilters==1.0.0
django-weasyprint==1.1.0.post1
djangorestframework==3.14.0
dulwich==0.20.50
et-xmlfile==1.1.0
filelock==3.8.1
firebase-admin==6.2.0
fonttools==4.38.0
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==21.2.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-metadata==5.1.0
jaraco.classes==3.2.3
jsonfield==3.1.0
jsonschema==4.17.3
keyring==23.11.0
lockfile==0.12.2
more-itertools==9.0.0
msgpack==1.0.4
openpyxl==3.1.2
packaging==21.3
pexpect==4.8.0
Pillow==9.3.0
pkginfo==1.9.2
platformdirs==2.5.4
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
proto-plus==1.22.3
protobuf==4.21.12
psycopg2==2.9.6
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.9.1
pycparser==2.21
pydyf==0.5.0
PyJWT==2.7.0
pylev==1.4.0
pyparsing==3.0.9
pyphen==0.13.1
pyrsistent==0.19.2
pytz==2022.7.1
reportlab==3.6.12
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
shellingham==1.5.0
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.17.0
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==6.5.0
xattr==0.9.9
zipp==3.11.0
zopfli==0.2.2
